requirements.txt
optuna~=3.0.5
mlflow~=2.1.1
torch~=1.12.0
torchvision~=0.13.0
albumentations~=1.3.0
pytorch_lightning~=1.5.10
pyyaml~=5.4.1
